Move adapterInstanceId related methods into hono-client-command.
