+++ clients/command/src/test/java/org/eclipse/hono/client/command/CommandRoutingUtilTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2022 Contributors to the Eclipse Foundation
+ * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -11,26 +11,27 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.util;
+package org.eclipse.hono.client.command;
 
 import static com.google.common.truth.Truth.assertThat;
 
+import org.eclipse.hono.util.Pair;
 import org.junit.jupiter.api.Test;
 
 
 /**
- * Tests verifying behavior of {@link CommandConstants}.
+ * Tests verifying behavior of {@link CommandRoutingUtil}.
  *
  */
-public class CommandConstantsTest {
+public class CommandRoutingUtilTest {
 
     @Test
     void testGetK8sPodNameAndContainerIdFromAdapterInstanceId() {
 
         final String podName = "myPodName";
         final String containerId = "012345678901";
-        final String newAdapterInstanceId = CommandConstants.getNewAdapterInstanceIdForK8sEnv(podName, containerId, 1);
-        final Pair<String, String> podNameAndContainerId = CommandConstants.getK8sPodNameAndContainerIdFromAdapterInstanceId(
+        final String newAdapterInstanceId = CommandRoutingUtil.getNewAdapterInstanceIdForK8sEnv(podName, containerId, 1);
+        final Pair<String, String> podNameAndContainerId = CommandRoutingUtil.getK8sPodNameAndContainerIdFromAdapterInstanceId(
                 newAdapterInstanceId);
         assertThat(podNameAndContainerId).isNotNull();
         assertThat(podNameAndContainerId.one()).isEqualTo(podName);
